some changes in updateMehtods

+++ src/main/java/com/hotelmangementsystem/application/service/BookedPriceServiceImpl.java
@@ -2,6 +2,7 @@ package com.hotelmangementsystem.application.service;
 
 import com.hotelmangementsystem.application.entity.BookedPrice;
 import com.hotelmangementsystem.application.repository.BookedPriceRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,7 +21,7 @@ public class BookedPriceServiceImpl implements BookedPriceService{
 
     @Override
     public BookedPrice getBookedPrice(Long id) {
-        return bookedPriceRepository.findById(id).get();
+        return bookedPriceRepository.findById(id).orElse(null);
     }
 
     @Override
@@ -30,13 +31,24 @@ public class BookedPriceServiceImpl implements BookedPriceService{
 
     @Override
     public BookedPrice updateBookedPrice(Long id, BookedPrice bookedPrice) {
-        BookedPrice bookedPriceDB = bookedPriceRepository.findById(id).get();
-        bookedPriceDB.setPrice(bookedPrice.getPrice());
-        return bookedPriceRepository.save(bookedPriceDB);
+        BookedPrice bookedPriceDB = bookedPriceRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("BookedPrice with id " + id + " does not exists"));
+
+        if(bookedPrice != null){
+            bookedPriceDB.setType(bookedPrice.getType());
+            bookedPriceDB.setName(bookedPrice.getName());
+            bookedPriceDB.setPrice(bookedPrice.getPrice());
+            return bookedPriceRepository.save(bookedPriceDB);
+        }
+        return null;
+
     }
 
     @Override
     public String deleteBookedPrice(Long id) {
+        boolean exists = bookedPriceRepository.existsById(id);
+        if(!exists) {
+            throw new EntityNotFoundException("BookedPrice with id " + id + " does not exists");
+        }
         bookedPriceRepository.deleteById(id);
         return "deleted";
     }
