+++ src/main/java/com/hotelmangementsystem/application/service/BookingServiceImpl.java
@@ -2,6 +2,7 @@ package com.hotelmangementsystem.application.service;
 
 import com.hotelmangementsystem.application.entity.Booking;
 import com.hotelmangementsystem.application.repository.BookingRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,7 +21,7 @@ public class BookingServiceImpl implements BookingService{
 
     @Override
     public Booking getBooking(Long id) {
-        return bookingRepository.findById(id).get();
+        return bookingRepository.findById(id).orElse(null);
     }
 
     @Override
@@ -30,13 +31,26 @@ public class BookingServiceImpl implements BookingService{
 
     @Override
     public Booking updateBooking(Long id, Booking booking) {
-        Booking bookingDB = bookingRepository.findById(id).get();
-        bookingDB.setStatus(booking.getStatus());
-        return bookingRepository.save(bookingDB);
+        Booking bookingDB = bookingRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Booking with id " + id + " does not exists"));
+
+        if(booking != null){
+            bookingDB.setCheckInDate(booking.getCheckInDate());
+            bookingDB.setCheckOutDate(booking.getCheckOutDate());
+            bookingDB.setNoOfAdult(booking.getNoOfAdult());
+            bookingDB.setStatus(booking.getStatus());
+            bookingDB.setTotal(booking.getTotal());
+            return bookingRepository.save(bookingDB);
+        }
+        return null;
+
     }
 
     @Override
     public String deleteBooking(Long id) {
+        boolean exists = bookingRepository.existsById(id);
+        if(!exists) {
+            throw new EntityNotFoundException("Booking with id " + id + " does not exists");
+        }
         bookingRepository.deleteById(id);
         return "deleted";
     }
