+++ src/main/java/com/hotelmangementsystem/application/service/DiscountServiceImpl.java
@@ -2,6 +2,7 @@ package com.hotelmangementsystem.application.service;
 
 import com.hotelmangementsystem.application.entity.Discount;
 import com.hotelmangementsystem.application.repository.DiscountRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,7 +21,7 @@ public class DiscountServiceImpl implements DiscountService{
 
     @Override
     public Discount getDiscount(Long id) {
-        return discountRepository.findById(id).get();
+        return discountRepository.findById(id).orElse(null);
     }
 
     @Override
@@ -30,13 +31,24 @@ public class DiscountServiceImpl implements DiscountService{
 
     @Override
     public Discount updateDiscount(Long id, Discount discount) {
-        Discount discountDB = discountRepository.findById(id).get();
-        discountDB.setPercentage(discount.getPercentage());
-        return discountRepository.save(discountDB);
+        Discount discountDB = discountRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Discount with id " + id + " does not exists"));
+
+        if(discount != null){
+            discountDB.setName(discount.getName());
+            discountDB.setDescription(discount.getDescription());
+            discountDB.setPercentage(discount.getPercentage());
+            return discountRepository.save(discountDB);
+        }
+        return null;
+
     }
 
     @Override
     public String deleteDiscount(Long id) {
+        boolean exists = discountRepository.existsById(id);
+        if(!exists) {
+            throw new EntityNotFoundException("Discount with id " + id + " does not exists");
+        }
         discountRepository.deleteById(id);
         return "deleted";
     }
