+++ src/main/java/com/hotelmangementsystem/application/service/PassengerServiceImpl.java
@@ -2,6 +2,7 @@ package com.hotelmangementsystem.application.service;
 
 import com.hotelmangementsystem.application.entity.Passenger;
 import com.hotelmangementsystem.application.repository.PassengerRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,7 +21,7 @@ public class PassengerServiceImpl implements PassengerService{
 
     @Override
     public Passenger getPassenger(Long id) {
-        return passengerRepository.findById(id).get();
+        return passengerRepository.findById(id).orElse(null);
     }
 
     @Override
@@ -30,13 +31,23 @@ public class PassengerServiceImpl implements PassengerService{
 
     @Override
     public Passenger updatePassenger(Long id, Passenger passenger) {
-        Passenger passengerDB = passengerRepository.findById(id).get();
-        passengerDB.setName(passenger.getName());
-        return passengerRepository.save(passengerDB);
+        Passenger passengerDB = passengerRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Passenger with id " + id + " does not exists"));
+
+        if(passenger != null){
+            passengerDB.setPhone(passenger.getPhone());
+            passengerDB.setName(passenger.getName());
+            return passengerRepository.save(passengerDB);
+        }
+        return null;
+
     }
 
     @Override
     public String deletePassenger(Long id) {
+        boolean exists = passengerRepository.existsById(id);
+        if(!exists){
+            throw new EntityNotFoundException("Passenger with id " + id + " does not exists");
+        }
         passengerRepository.deleteById(id);
         return "deleted";
     }
