+++ src/main/java/com/hotelmangementsystem/application/service/PaymentServiceImpl.java
@@ -2,6 +2,7 @@ package com.hotelmangementsystem.application.service;
 
 import com.hotelmangementsystem.application.entity.Payment;
 import com.hotelmangementsystem.application.repository.PaymentRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,7 +21,7 @@ public class PaymentServiceImpl implements PaymentService{
 
     @Override
     public Payment getPayment(Long id) {
-        return paymentRepository.findById(id).get();
+        return paymentRepository.findById(id).orElse(null);
     }
 
     @Override
@@ -30,13 +31,24 @@ public class PaymentServiceImpl implements PaymentService{
 
     @Override
     public Payment updatePayment(Long id, Payment payment) {
-        Payment paymentDB = paymentRepository.findById(id).get();
-        paymentDB.setAmount(payment.getAmount());
-        return paymentRepository.save(paymentDB);
+        Payment paymentDB = paymentRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Payment with id " + id + " does not exists"));
+
+        if(payment != null){
+            paymentDB.setDate(payment.getDate());
+            paymentDB.setAmount(payment.getAmount());
+            paymentDB.setMethod(payment.getMethod());
+            return paymentRepository.save(paymentDB);
+        }
+        return null;
+
     }
 
     @Override
     public String deletePayment(Long id) {
+        boolean exists = paymentRepository.existsById(id);
+        if(!exists){
+            throw new EntityNotFoundException("Payment with id " + id + " does not exists");
+        }
         paymentRepository.deleteById(id);
         return "deleted";
     }
